fixed compilation nits on windows
git-svn-id: https://commonmusic.svn.sourceforge.net/svnroot/commonmusic/trunk@1952 d60aafaf-7936-0410-ba4d-d307febf7868

+++ src/MidiPlaybackThread.h
@@ -313,7 +313,7 @@ class MidiPlaybackThread : public juce::Thread
   {
     juce::ScopedLock mylock(pblock);
     if (tpb<1) tpb=1;
-    ticks=tpb;
+    ticks=(int)tpb;
     position.tick=1.0/tpb;
   }
